+++ sdk/core/azure_core/src/credentials.rs
@@ -100,7 +100,4 @@ impl AccessToken {
 pub trait TokenCredential: Send + Sync + Debug {
     /// Gets a `AccessToken` for the specified resource
     async fn get_token(&self, scopes: &[&str]) -> crate::Result<AccessToken>;
-
-    /// Clear the credential's cache.
-    async fn clear_cache(&self) -> crate::Result<()>;
 }