+++ sdk/identity/azure_identity/src/credentials/azure_cli_credentials.rs
@@ -293,11 +293,6 @@ impl TokenCredential for AzureCliCredential {
     async fn get_token(&self, scopes: &[&str]) -> azure_core::Result<AccessToken> {
         self.cache.get_token(scopes, self.get_token(scopes)).await
     }
-
-    /// Clear the credential's cache.
-    async fn clear_cache(&self) -> azure_core::Result<()> {
-        self.cache.clear().await
-    }
 }
 
 impl From<TokenCredentialOptions> for AzureCliCredentialOptions {