+++ sdk/identity/azure_identity/src/credentials/workload_identity_credentials.rs
@@ -107,10 +107,6 @@ impl TokenCredential for WorkloadIdentityCredential {
     async fn get_token(&self, scopes: &[&str]) -> azure_core::Result<AccessToken> {
         self.0.get_token(scopes).await
     }
-
-    async fn clear_cache(&self) -> azure_core::Result<()> {
-        self.0.clear_cache().await
-    }
 }
 
 // TODO: Should probably remove this once we consolidate and unify credentials.